modify strncmp

+++ Libft/ft_strncmp.c
@@ -22,7 +22,7 @@ int	ft_strncmp(const char *str1, const char *str2, size_t n)
 		if (!*(str1 + index) && !*(str2 + index))
 			return (0);
 		if (*(str1 + index) != *(str2 + index))
-			return (*(str1 + index) - *(str2 + index));
+			return ((unsigned int)*(str1 + index) - (unsigned int)*(str2 + index));
 		index++;
 	}
 	return (0);
